fix: refactord intern dashboard
